requirements.txt
aiohttp==3.8.1
aiohttp-swagger==1.0.16
aiosignal==1.2.0
apispec==3.3.2
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
elastic-transport==8.1.2
elasticsearch==8.1.3
frozenlist==1.3.0
idna==3.3
iniconfig==1.1.1
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.15.0
mccabe==0.6.1
mirakuru==2.4.2
motor==3.0.0
multidict==6.0.2
packaging==21.3
pluggy==1.0.0
port-for==0.6.2
psutil==5.9.0
py==1.11.0
pymongo==4.1.1
pyparsing==3.0.8
pytest==7.1.2
pytest-asyncio==0.18.3
python-dotenv==0.20.0
PyYAML==6.0
tomli==2.0.1
urllib3==1.26.9
waitress==2.1.1
webargs==5.5.3
yarl==1.7.2
